+++ .babelrc.js
@@ -10,9 +10,6 @@ module.exports = (api) => {
 					namespace: 'sc-',
 				},
 			],
-			// "react-hot-loader/babel"
-			// '@babel/plugin-transform-destructuring',
-			// ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
 			'@babel/plugin-transform-runtime',
 		],
 		presets: [