+++ src/components/MetadataForm/MetadataForm.test.tsx
@@ -1,16 +1,17 @@
 import React, { ComponentProps } from 'react';
 import renderWithTheme from '~/tests/utils/withTheme';
-import MetadataForm from './MetadataForm';
+import MetadataForm, { Props } from './MetadataForm';
 import { schema, uischema } from './__stubs__';
 import { screen } from '@testing-library/react';
 import { JsonForms } from '@jsonforms/react';
 import MetdadataColumnHeaders from '../MetadataColumnHeaders/MetadataColumnHeaders';
 import Form from '../Form/Form';
+import { mockComponentProps, mocked } from '~/tests/helpers';
 
 jest.mock('../MetadataColumnHeaders/MetadataColumnHeaders');
 jest.mock('../Form/Form');
 
-const props: ComponentProps<typeof MetadataForm> = {
+const props: Props = {
 	ajv: undefined,
 	schema,
 	uischema,
@@ -19,6 +20,8 @@ const props: ComponentProps<typeof MetadataForm> = {
 	validationMode: 'ValidateAndShow',
 	onChange: jest.fn(),
 };
+
+const MetdadataColumnHeadersMock = mocked(MetdadataColumnHeaders);
 
 describe('<MetadataForm />', () => {
 	const render = (customProps: Partial<ComponentProps<typeof JsonForms>> = {}) => {
@@ -33,15 +36,12 @@ describe('<MetadataForm />', () => {
 
 	test('Renders column headers', () => {
 		render();
-		expect(MetdadataColumnHeaders).toHaveBeenCalledWith(
-			{
-				columns: [
-					{ header: 'Metadataveld', width: 50 },
-					{ header: 'Waarde', width: 50 },
-				],
-			},
-			{},
-		);
+		expect(mockComponentProps<ComponentProps<typeof MetdadataColumnHeaders>>(MetdadataColumnHeadersMock)).toEqual({
+			columns: [
+				{ header: 'Metadataveld', width: 50 },
+				{ header: 'Waarde', width: 50 },
+			],
+		});
 	});
 
 	test('Renders form', () => {