+++ src/components/MetadataForm/__stubs__/uischema.ts
@@ -1,6 +1,13 @@
-export const uischema = {
+import { RowLayoutSchema } from '../../../types';
+
+export const uischema: RowLayoutSchema = {
 	type: 'RowLayout',
 	elements: [
+		{
+			type: 'Control',
+			label: 'Textfield',
+			scope: '#/properties/textField',
+		},
 		{
 			type: 'Control',
 			label: 'Documentomschrijving',