+++ src/components/Wizard/WizardStyles.ts
@@ -1,5 +1,6 @@
 import { Button, Heading, themeSpacing } from '@amsterdam/asc-ui';
 import styled from 'styled-components';
+import { ComponentProps } from 'react';
 
 export const PreviousButtonStyle = styled(Button)`
 	padding-right: 20px;
@@ -15,6 +16,8 @@ export const ModalContentStyle = styled.div`
 	padding: ${themeSpacing(8)};
 `;
 
-export const ModalTopBarStyle = styled(Heading)`
+export const ModalTopBarStyle = styled(Heading)<ComponentProps<typeof Heading>>`
 	font-size: 18px;
+	line-height: 25px;
+	font-weight: 700;
 `;