+++ src/components/Wizard/__mocks__/Step2.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Step2 = jest.fn().mockImplementation(() => <div data-testid="step-2" />);
-
-export default Step2;