+++ CHANGELOG.md
@@ -2,6 +2,139 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.3-alpha.2](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.3-alpha.1...v0.0.3-alpha.2) (2021-12-31)
+
+
+### Bug Fixes
+
+* **creatable:** TypeScript errors no more ([58604b2](https://github.com/Amsterdam/bmi-dms-upload/commit/58604b26db38cb4fb1ef1dbcfab7582d5b2ca733))
+
+### [0.0.3-alpha.1](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.3-alpha.0...v0.0.3-alpha.1) (2021-12-30)
+
+
+### Features
+
+* **bulkmetadata:** add StaticMetadata component ([c0b4551](https://github.com/Amsterdam/bmi-dms-upload/commit/c0b455102c0e8a405361eb3436c64dac3462fc40))
+* **bulkmetadata:** add StaticMetadata component ([c4af604](https://github.com/Amsterdam/bmi-dms-upload/commit/c4af604f27c8a25f59379a7608ed949da1d91e2d))
+* **bulkmetadata:** add StaticMetadata tests ([0c1d459](https://github.com/Amsterdam/bmi-dms-upload/commit/0c1d459c0ab7d285b7976f98ce2e25647f8ea99e))
+* **bulkmetadata:** add StaticMetadata tests ([fc805cb](https://github.com/Amsterdam/bmi-dms-upload/commit/fc805cbb3a483e1c1a41c0d64927b8489a3a1626))
+* **bulkmetadata:** update test ([8263a7a](https://github.com/Amsterdam/bmi-dms-upload/commit/8263a7ab1d5c6b3ac7bf8bc760103fe6b909250a))
+* **bulkmetadata:** updated by review comments ([40c00e4](https://github.com/Amsterdam/bmi-dms-upload/commit/40c00e451ceb87d5177130e112bb65cdc5c35322))
+* **bulkmetadata:** updated by review comments ([dfb4a1f](https://github.com/Amsterdam/bmi-dms-upload/commit/dfb4a1fcb77b7a4b53515bbfdaaa2c9f91afd750))
+* **bulkmetadata:** updated by review comments ([40eab59](https://github.com/Amsterdam/bmi-dms-upload/commit/40eab59524b84f7a3206c6e2c6a3df8cbbd5d30e))
+* **bulkmetadata:** updated by review comments ([44ad9b1](https://github.com/Amsterdam/bmi-dms-upload/commit/44ad9b1c8d6d724bdf423af17bfc0bec4041ce19))
+* **bulkmetadata:** updated by review comments ([ccc0d6b](https://github.com/Amsterdam/bmi-dms-upload/commit/ccc0d6b7ca03a4ab66c30ec186f184385ed35644))
+* **step1:** placeholder and maxFiles for step1 are now configurable by prop ([8471ef4](https://github.com/Amsterdam/bmi-dms-upload/commit/8471ef48535f60b4c680243bd05a52d0631a6d17))
+
+
+### Bug Fixes
+
+* **header:** styling in line with design system ([64d1461](https://github.com/Amsterdam/bmi-dms-upload/commit/64d14617a04718ef2e758431253583412aa0e304))
+* **metadataform:** fixed alignment of label with text field ([a54dd80](https://github.com/Amsterdam/bmi-dms-upload/commit/a54dd809e2bc6115961a6443bbe195ff989c0ac6))
+* **styling:** styling fixes after design QA ([2d5e62e](https://github.com/Amsterdam/bmi-dms-upload/commit/2d5e62e902f971f290d49f217c640eed8fb0388c))
+* **wizard:** tweaking buttons in footer after design QA ([5af866d](https://github.com/Amsterdam/bmi-dms-upload/commit/5af866d44a0833d82240ad1922fe6d88f8228c76))
+* **wizard:** tweaking buttons in footer after design QA ([3fd67ab](https://github.com/Amsterdam/bmi-dms-upload/commit/3fd67abe459b330064fa9bf4d97ba97d510c0093))
+
+### [0.0.3-alpha.0](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.2...v0.0.3-alpha.0) (2021-12-17)
+
+
+### Features
+
+* **dms-upload:** bumped component library version and adjusted storybook, readme ([7b08195](https://github.com/Amsterdam/bmi-dms-upload/commit/7b081957047bd4ef76490785b0b64cf52d8d3534))
+
+### [0.0.2](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1...v0.0.2) (2021-12-15)
+
+
+### Features
+
+* **bulkmetadata:** add StaticMetadata component ([47d03df](https://github.com/Amsterdam/bmi-dms-upload/commit/47d03df514b73dc46d5e641868d0481c5be53635))
+* **bulkmetadata:** add StaticMetadata tests ([c49b413](https://github.com/Amsterdam/bmi-dms-upload/commit/c49b413ebdefdb53e4d843e086a0abfe014fbd7a))
+* **bulkmetadata:** update test ([533ba3c](https://github.com/Amsterdam/bmi-dms-upload/commit/533ba3c76b3b9f21534af3e3eba621be9caea7e3))
+* **bulkmetadata:** updated by review comments ([81c0892](https://github.com/Amsterdam/bmi-dms-upload/commit/81c08920e24cb052149efe12137c7b9494103cca))
+* **bulkmetadata:** updated by review comments ([5bbc734](https://github.com/Amsterdam/bmi-dms-upload/commit/5bbc734a4c74c6503f071007c6bcda15c96528a2))
+* **bulkmetadata:** updated by review comments ([a9c73e8](https://github.com/Amsterdam/bmi-dms-upload/commit/a9c73e825ab498660f6683086e024975bb725c85))
+* **bulkmetadata:** updated by review comments ([1a1fa9e](https://github.com/Amsterdam/bmi-dms-upload/commit/1a1fa9e7d8c3d3ccabe242339a4a7891c31e3f57))
+
+
+### Bug Fixes
+
+* **modal:** freshly squeezed package-lock.json ([08f5167](https://github.com/Amsterdam/bmi-dms-upload/commit/08f5167f46a2be53def251a4c0b2356dfb0d47e7))
+* **modal:** modal actions were not split left and right in the production build ([d8a270f](https://github.com/Amsterdam/bmi-dms-upload/commit/d8a270fa280e4cfeadd30cac8d355471affb90e9))
+
+### [0.0.2-alpha.0](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.10...v0.0.2-alpha.0) (2021-12-13)
+
+
+### Features
+
+* **bulkmetadata:** add StaticMetadata component ([47d03df](https://github.com/Amsterdam/bmi-dms-upload/commit/47d03df514b73dc46d5e641868d0481c5be53635))
+* **bulkmetadata:** add StaticMetadata tests ([c49b413](https://github.com/Amsterdam/bmi-dms-upload/commit/c49b413ebdefdb53e4d843e086a0abfe014fbd7a))
+* **bulkmetadata:** update test ([533ba3c](https://github.com/Amsterdam/bmi-dms-upload/commit/533ba3c76b3b9f21534af3e3eba621be9caea7e3))
+* **bulkmetadata:** updated by review comments ([81c0892](https://github.com/Amsterdam/bmi-dms-upload/commit/81c08920e24cb052149efe12137c7b9494103cca))
+* **bulkmetadata:** updated by review comments ([5bbc734](https://github.com/Amsterdam/bmi-dms-upload/commit/5bbc734a4c74c6503f071007c6bcda15c96528a2))
+* **bulkmetadata:** updated by review comments ([a9c73e8](https://github.com/Amsterdam/bmi-dms-upload/commit/a9c73e825ab498660f6683086e024975bb725c85))
+* **bulkmetadata:** updated by review comments ([1a1fa9e](https://github.com/Amsterdam/bmi-dms-upload/commit/1a1fa9e7d8c3d3ccabe242339a4a7891c31e3f57))
+
+
+### Bug Fixes
+
+* **header:** styling in line with design system ([36ef717](https://github.com/Amsterdam/bmi-dms-upload/commit/36ef717cbc1ab3b05e28306e41ed1011f71b8695))
+* **metadataform:** fixed alignment of label with text field ([309fc33](https://github.com/Amsterdam/bmi-dms-upload/commit/309fc33c392188892f6ef671efd071cc1e99cc32))
+* **styling:** styling fixes after design QA ([8cf9698](https://github.com/Amsterdam/bmi-dms-upload/commit/8cf9698e55dd6c43fa5d947011fc831cb7276be0))
+* **wizard:** tweaking buttons in footer after design QA ([7948562](https://github.com/Amsterdam/bmi-dms-upload/commit/794856216d044221b6daa594fb4f5ceddcf8f42f))
+
+### [0.0.1](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.10...v0.0.1) (2021-12-07)
+
+
+### Bug Fixes
+
+* **header:** styling in line with design system ([36ef717](https://github.com/Amsterdam/bmi-dms-upload/commit/36ef717cbc1ab3b05e28306e41ed1011f71b8695))
+* **metadataform:** fixed alignment of label with text field ([309fc33](https://github.com/Amsterdam/bmi-dms-upload/commit/309fc33c392188892f6ef671efd071cc1e99cc32))
+* **styling:** styling fixes after design QA ([8cf9698](https://github.com/Amsterdam/bmi-dms-upload/commit/8cf9698e55dd6c43fa5d947011fc831cb7276be0))
+* **wizard:** tweaking buttons in footer after design QA ([7948562](https://github.com/Amsterdam/bmi-dms-upload/commit/794856216d044221b6daa594fb4f5ceddcf8f42f))
+
+### [0.0.1-alpha.10](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.9...v0.0.1-alpha.10) (2021-12-07)
+
+
+### Features
+
+* **step1:** placeholder and maxFiles for step1 are now configurable by prop ([c3646a6](https://github.com/Amsterdam/bmi-dms-upload/commit/c3646a6b4597f9c66953229c47853278898ed606))
+
+### [0.0.1-alpha.9](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.8...v0.0.1-alpha.9) (2021-11-25)
+
+
+### Bug Fixes
+
+* **adddocumentbutton:** changed button vairant ([304c0ff](https://github.com/Amsterdam/bmi-dms-upload/commit/304c0ff20b71d3788f6de08bcc6ded80e28c29f4))
+
+### [0.0.1-alpha.8](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.7...v0.0.1-alpha.8) (2021-11-24)
+
+
+### Bug Fixes
+
+* **step1:** file upload should only accept a single file ([53f3bb8](https://github.com/Amsterdam/bmi-dms-upload/commit/53f3bb812a9cfbebf90836357c437e910f47284e))
+* **wizard:** bind cancel procedure to close button in modal header ([17c9e04](https://github.com/Amsterdam/bmi-dms-upload/commit/17c9e04814be0b3202a8344f8d11f20b52af8d04))
+
+### [0.0.1-alpha.7](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.6...v0.0.1-alpha.7) (2021-11-22)
+
+
+### Features
+
+* **types:** exporting json schema types ([09d38ed](https://github.com/Amsterdam/bmi-dms-upload/commit/09d38ed9fcaaed4743db2e3b9b84989fb1e76286))
+
+### [0.0.1-alpha.6](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.5...v0.0.1-alpha.6) (2021-11-18)
+
+
+### Features
+
+* **metadatacolumnheaders:** added MetadataColumnHeaders component + tests + storybook entry ([628b86b](https://github.com/Amsterdam/bmi-dms-upload/commit/628b86b465a5a2005007ac3027e6bca0de7b6639))
+* **validation:** wip commit for validation logic with latest alpha release of jsonforms ([d78e55e](https://github.com/Amsterdam/bmi-dms-upload/commit/d78e55e08e50e9498516bf74afbfcbcf8e310f01))
+
+
+### Bug Fixes
+
+* **tsconfig:** ignoring **/__mocks/* for TS build purposes ([fdc3383](https://github.com/Amsterdam/bmi-dms-upload/commit/fdc33833887fc3ea225427f5ae32f1d8293df91b))
+* **tsconfig:** reverted esModuleInterop back to true while fixing source of the problem ([eeb3acd](https://github.com/Amsterdam/bmi-dms-upload/commit/eeb3acd3e881a56564e25730b3485d943dfc2fa4))
+
 ### [0.0.1-alpha.5](https://github.com/Amsterdam/bmi-dms-upload/compare/v0.0.1-alpha.4...v0.0.1-alpha.5) (2021-10-27)